+++ Shared/ViewModels/AdminDashboard/DeviceDetailViewModel.swift
