+++ Swiftfin/Views/AdminDashboardView/ActiveSessionDetailView/Components/StreamSection.swift
