+++ Swiftfin/Views/AdminDashboardView/AddTaskTriggerView/Components/IntervalRow.swift
