+++ Swiftfin/Views/AdminDashboardView/Components/DeviceSection.swift
@@ -9,7 +9,7 @@
 import JellyfinAPI
 import SwiftUI
 
-extension UserDashboardView {
+extension AdminDashboardView {
 
     struct DeviceSection: View {
