+++ Swiftfin/Views/AdminDashboardView/DevicesView/Components/DeviceRow.swift
