+++ Swiftfin/Views/AdminDashboardView/EditServerTaskView/Components/Sections/TriggersSection.swift
@@ -14,7 +14,7 @@ extension EditServerTaskView {
     struct TriggersSection: View {
 
         @EnvironmentObject
-        private var router: SettingsCoordinator.Router
+        private var router: AdminDashboardCoordinator.Router
 
         @ObservedObject
         var observer: ServerTaskObserver
