+++ Swiftfin/Views/AdminDashboardView/EditServerTaskView/Components/TriggerRow.swift
