+++ Swiftfin/Views/SettingsView/SettingsView/SettingsView.swift
@@ -46,7 +46,7 @@ struct SettingsView: View {
                 if viewModel.userSession.user.isAdministrator {
                     ChevronButton(L10n.dashboard)
                         .onSelect {
-                            router.route(to: \.userDashboard)
+                            router.route(to: \.adminDashboard)
                         }
                 }
             }
